requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
flake8==6.0.0
flake8-docstrings==1.7.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
mccabe==0.7.0
pycodestyle==2.10.0
pydocstyle==6.3.0
pyflakes==3.0.1
python-dotenv==0.21.1
python-telegram-bot==20.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
snowballstemmer==2.2.0
urllib3==1.26.14
